+++ R/run_checks.R
@@ -143,6 +143,15 @@ run_congruence_checks <- function(directory = here::here(),
              warn_count <<- warn_count + 1
              cli::cli_bullets(c(w$message, w$body))
            })
+  tryCatch(test_content_units(metadata),
+           error = function(e) {
+             err_count <<- err_count + 1
+             cli::cli_bullets(c(e$message, e$body))
+           },
+           warning = function(w) {
+             warn_count <<- warn_count + 1
+             cli::cli_bullets(c(w$message, w$body))
+           })
   tryCatch(test_doi(metadata),
            error = function(e) {
              err_count <<- err_count + 1